requirements.txt
streamlit
streamlit-webrtc
numpy
opencv-python-headless
tensorflow-cpu
mediapipe
scikit-learn